refactor(Core): cache stmt results to avoid duplicate query in recursive list_dids
Closes #8478
